+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+### Added
+
+- `imageUrl`, `alwaysSubmitViaPendingQueue`, `taskAllocations`, `taskGroupInstanceAllocations`, `slug`, `allowAdhocTasks`, `listDisplayAttributes` to `FormsAppMenuItem`
+
+### Breaking
+
+- Renamed class `FormsAppMenuItems` to `FormsAppMenuItem`
+
 ## [11.3.2] - 2026-05-07
 
 ### Added